requirements.txt
bibtexparser==0.6.2
bottle==0.12.13
cachetools==2.0.0
certifi==2017.4.17
chardet==3.0.3
firebase-admin==2.0.0
gcloud==0.17.0
google-auth==1.0.1
googleapis-common-protos==1.5.2
httplib2==0.10.3
idna==2.5
Jinja2==2.9.6
jws==0.1.3
lxml==3.8.0
Markdown==2.6.8
MarkupSafe==1.0
oauth2client==3.0.0
Paste==2.0.3
pkgconfig==1.2.2
protobuf==3.3.0
pyaml==16.12.2
pyasn1==0.2.3
pyasn1-modules==0.0.9
pycryptodome==3.4.3
pymongo==3.4.0
python-jwt==2.0.1
PyYAML==3.12
requests==2.17.3
requests-toolbelt==0.7.0
rsa==3.4.2
six==1.10.0
urllib3==1.21.1
